CT feat: stat

+++ client/AgileOnline/Assets/@Scripts/Controllers/Skill/EliteMonsterProjectileController.cs
@@ -15,7 +15,7 @@ public class EliteMonsterProjectileController : SkillController
     public string Description { get; private set; }
     public string IconName { get; private set; }
     public float CoolTime { get; private set; }
-    public int Damage { get; private set; }
+    public float Damage { get; private set; }
     public float LifeTime { get; private set; } = 10;
     public float Speed { get; private set; }
     public int ProjectileNum { get; private set; }
@@ -73,7 +73,7 @@ public class EliteMonsterProjectileController : SkillController
         if (player.IsValid() == false)
             return;
         
-        int damage = Damage;
+        float damage = Damage;
         player.OnDamaged(_owner, ref damage);
         
         Managers.Object.Despawn(this);
